+++ bus-health/src/main/java/org/miaixz/bus/health/unix/aix/software/AixOperatingSystem.java
@@ -284,7 +284,8 @@ public class AixOperatingSystem extends AbstractOperatingSystem {
      */
     @Override
     public int getThreadId() {
-        return AixLibc.INSTANCE.thread_self();
+        // OSThread.getThreadId() is int, so narrow at the API boundary.
+        return AixLibc.INSTANCE.thread_self().intValue();
     }
 
     /**
